[NU-2152] Missing parameters in node are only reported as warnings in logs - no validation error is returned
