+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/DynamicNodeValidator.scala
@@ -195,8 +195,7 @@ class DynamicNodeValidator(
             validatorsCompilationResult.andThen { validators =>
               expressionCompiler
                 .compileBranchParam(branchParams, branchContexts, parameter)
-                .map((_, None))
-                .andThen(Validations.validate(validators, _))
+                .andThen(typedParam => Validations.validate(validators, typedParam).map(_ => (typedParam, None)))
             }
           }
       } else {
@@ -220,8 +219,9 @@ class DynamicNodeValidator(
         validatorsCompilationResult.andThen { validators =>
           expressionCompiler
             .compileParam(singleParam, ctxToUse, parameter)
-            .map((_, extraNodeParamOpt))
-            .andThen(Validations.validate(validators, _))
+            .andThen(typedParam =>
+              Validations.validate(validators, typedParam).map(_ => (typedParam, extraNodeParamOpt))
+            )
         }
       }
     }
